+++ docker/resource_docker_container_test.go
@@ -420,15 +420,84 @@ func TestAccDockerContainer_port_internal(t *testing.T) {
 					testCheck,
 					resource.TestCheckResourceAttr("docker_container.foo", "name", "tf-test"),
 					resource.TestCheckResourceAttr("docker_container.foo", "ports.#", "1"),
-					resource.TestCheckResourceAttr("docker_container.foo", "ports.2978131916.internal", "80"),
-					resource.TestCheckResourceAttr("docker_container.foo", "ports.2978131916.ip", "0.0.0.0"),
-					resource.TestCheckResourceAttr("docker_container.foo", "ports.2978131916.protocol", "tcp"),
-					resource.TestCheckResourceAttr("docker_container.foo", "ports.2978131916.external", "32678"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.0.internal", "80"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.0.ip", "0.0.0.0"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.0.protocol", "tcp"),
+					testValueHigherEqualThan("docker_container.foo", "ports.0.external", 32768),
 				),
 			},
 		},
 	})
 }
+func TestAccDockerContainer_port_multiple_internal(t *testing.T) {
+	var c types.ContainerJSON
+
+	testCheck := func(*terraform.State) error {
+		portMap := c.NetworkSettings.NetworkSettingsBase.Ports
+		portBindings, ok := portMap["80/tcp"]
+		if !ok || len(portMap["80/tcp"]) == 0 {
+			return fmt.Errorf("Port 80 on tcp is not set")
+		}
+
+		portBindingsLength := len(portBindings)
+		if portBindingsLength != 1 {
+			return fmt.Errorf("Expected 1 binding on port 80, but was %d", portBindingsLength)
+		}
+
+		if len(portBindings[0].HostIP) == 0 {
+			return fmt.Errorf("Expected host IP to be set, but was empty")
+		}
+
+		if len(portBindings[0].HostPort) == 0 {
+			return fmt.Errorf("Expected host port to be set, but was empty")
+		}
+
+		portBindings, ok = portMap["81/tcp"]
+		if !ok || len(portMap["81/tcp"]) == 0 {
+			return fmt.Errorf("Port 81 on tcp is not set")
+		}
+
+		portBindingsLength = len(portBindings)
+		if portBindingsLength != 1 {
+			return fmt.Errorf("Expected 1 binding on port 81, but was %d", portBindingsLength)
+		}
+
+		if len(portBindings[0].HostIP) == 0 {
+			return fmt.Errorf("Expected host IP to be set, but was empty")
+		}
+
+		if len(portBindings[0].HostPort) == 0 {
+			return fmt.Errorf("Expected host port to be set, but was empty")
+		}
+
+		return nil
+	}
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:  func() { testAccPreCheck(t) },
+		Providers: testAccProviders,
+		Steps: []resource.TestStep{
+			resource.TestStep{
+				Config: testAccDockerContainerMultipleInternalPortConfig,
+				Check: resource.ComposeTestCheckFunc(
+					testAccContainerRunning("docker_container.foo", &c),
+					testCheck,
+					resource.TestCheckResourceAttr("docker_container.foo", "name", "tf-test"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.#", "2"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.0.internal", "80"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.0.ip", "0.0.0.0"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.0.protocol", "tcp"),
+					testValueHigherEqualThan("docker_container.foo", "ports.0.external", 32768),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.1.internal", "81"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.1.ip", "0.0.0.0"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.1.protocol", "tcp"),
+					testValueHigherEqualThan("docker_container.foo", "ports.1.external", 32768),
+				),
+			},
+		},
+	})
+}
+
 func TestAccDockerContainer_port(t *testing.T) {
 	var c types.ContainerJSON
 
@@ -466,10 +535,78 @@ func TestAccDockerContainer_port(t *testing.T) {
 					testCheck,
 					resource.TestCheckResourceAttr("docker_container.foo", "name", "tf-test"),
 					resource.TestCheckResourceAttr("docker_container.foo", "ports.#", "1"),
-					resource.TestCheckResourceAttr("docker_container.foo", "ports.2498386340.internal", "80"),
-					resource.TestCheckResourceAttr("docker_container.foo", "ports.2498386340.ip", "0.0.0.0"),
-					resource.TestCheckResourceAttr("docker_container.foo", "ports.2498386340.protocol", "tcp"),
-					resource.TestCheckResourceAttr("docker_container.foo", "ports.2498386340.external", "32787"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.0.internal", "80"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.0.ip", "0.0.0.0"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.0.protocol", "tcp"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.0.external", "32787"),
+				),
+			},
+		},
+	})
+}
+func TestAccDockerContainer_multiple_ports(t *testing.T) {
+	var c types.ContainerJSON
+
+	testCheck := func(*terraform.State) error {
+		portMap := c.NetworkSettings.NetworkSettingsBase.Ports
+		portBindings, ok := portMap["80/tcp"]
+		if !ok || len(portMap["80/tcp"]) == 0 {
+			return fmt.Errorf("Port 80 on tcp is not set")
+		}
+
+		portBindingsLength := len(portBindings)
+		if portBindingsLength != 1 {
+			return fmt.Errorf("Expected 1 binding on port 80, but was %d", portBindingsLength)
+		}
+
+		if len(portBindings[0].HostIP) == 0 {
+			return fmt.Errorf("Expected host IP to be set, but was empty")
+		}
+
+		if len(portBindings[0].HostPort) == 0 {
+			return fmt.Errorf("Expected host port to be set, but was empty")
+		}
+
+		portBindings, ok = portMap["81/tcp"]
+		if !ok || len(portMap["81/tcp"]) == 0 {
+			return fmt.Errorf("Port 81 on tcp is not set")
+		}
+
+		portBindingsLength = len(portBindings)
+		if portBindingsLength != 1 {
+			return fmt.Errorf("Expected 1 binding on port 81, but was %d", portBindingsLength)
+		}
+
+		if len(portBindings[0].HostIP) == 0 {
+			return fmt.Errorf("Expected host IP to be set, but was empty")
+		}
+
+		if len(portBindings[0].HostPort) == 0 {
+			return fmt.Errorf("Expected host port to be set, but was empty")
+		}
+
+		return nil
+	}
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:  func() { testAccPreCheck(t) },
+		Providers: testAccProviders,
+		Steps: []resource.TestStep{
+			resource.TestStep{
+				Config: testAccDockerContainerMultiplePortConfig,
+				Check: resource.ComposeTestCheckFunc(
+					testAccContainerRunning("docker_container.foo", &c),
+					testCheck,
+					resource.TestCheckResourceAttr("docker_container.foo", "name", "tf-test"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.#", "2"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.0.internal", "80"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.0.ip", "0.0.0.0"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.0.protocol", "tcp"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.0.external", "32787"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.1.internal", "81"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.1.ip", "0.0.0.0"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.1.protocol", "tcp"),
+					resource.TestCheckResourceAttr("docker_container.foo", "ports.1.external", "32788"),
 				),
 			},
 		},
@@ -505,6 +642,37 @@ func testAccContainerRunning(n string, container *types.ContainerJSON) resource.
 		}
 
 		return fmt.Errorf("Container not found: %s", rs.Primary.ID)
+	}
+}
+
+func testValueHigherEqualThan(name, key string, value int) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		ms := s.RootModule()
+		rs, ok := ms.Resources[name]
+		if !ok {
+			return fmt.Errorf("Not found: %s", name)
+		}
+
+		is := rs.Primary
+		if is == nil {
+			return fmt.Errorf("No primary instance: %s", name)
+		}
+
+		vRaw, ok := is.Attributes[key]
+		if !ok {
+			return fmt.Errorf("%s: Attribute '%s' not found", name, key)
+		}
+
+		v, err := strconv.Atoi(vRaw)
+		if err != nil {
+			return fmt.Errorf("'%s' is not a number", vRaw)
+		}
+
+		if v < value {
+			return fmt.Errorf("'%v' is smaller than '%v', but was expected to be equal or greater", v, value)
+		}
+
+		return nil
 	}
 }
 
@@ -658,6 +826,27 @@ resource "docker_container" "foo" {
 	}
 }
 `
+
+const testAccDockerContainerMultipleInternalPortConfig = `
+resource "docker_image" "foo" {
+	name = "nginx:latest"
+	keep_locally = true
+}
+
+resource "docker_container" "foo" {
+	name = "tf-test"
+	image = "${docker_image.foo.latest}"
+	
+	ports = [
+		{
+			internal = "80"
+		},
+		{
+			internal = "81"
+		}
+	]
+}
+`
 const testAccDockerContainerPortConfig = `
 resource "docker_image" "foo" {
 	name = "nginx:latest"
@@ -672,5 +861,27 @@ resource "docker_container" "foo" {
 		internal = "80"
 		external = "32787"
 	}
+}
+`
+const testAccDockerContainerMultiplePortConfig = `
+resource "docker_image" "foo" {
+	name = "nginx:latest"
+	keep_locally = true
+}
+
+resource "docker_container" "foo" {
+	name = "tf-test"
+	image = "${docker_image.foo.latest}"
+
+	ports = [
+		{
+			internal = "80"
+			external = "32787"
+		},
+		{
+			internal = "81"
+			external = "32788"
+		}
+	] 
 }
 `